fix: only fp16 evaluation is supported when ampO2 is enabled

+++ ppcls/engine/evaluation/classification.py
@@ -53,7 +53,7 @@ def classification_eval(engine, epoch_id=0):
             ]
         time_info["reader_cost"].update(time.time() - tic)
         batch_size = batch[0].shape[0]
-        batch[0] = paddle.to_tensor(batch[0]).astype("float32")
+        batch[0] = paddle.to_tensor(batch[0])
         if not engine.config["Global"].get("use_multilabel", False):
             batch[1] = batch[1].reshape([-1, 1]).astype("int64")
